% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruuvifunctions.R
\name{LoadRuuviExports}
\alias{LoadRuuviExports}
\title{read ruuvi data from local machine}
\usage{
LoadRuuviExports(path)
}
\arguments{
\item{path}{local path to csv files}
}
\description{
csv data to R
}
\examples{
 LoadRuuviExports(path = 'C:\\\\Users\\\\talon_000\\\\Dropbox\\\\2021\\\\ruuvi\\\\m210801saunakahala')
}
