% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruuvifunctions.R
\name{MakeDataWide}
\alias{MakeDataWide}
\title{MakeDataWide}
\usage{
MakeDataWide(data, timewindow_min = 5, tag_name = FALSE)
}
\arguments{
\item{pathRuuvitag}{local path to csv files}
}
\description{
data wrangling
}
\examples{
 \dontrun{
   data.wide <- MakeDataWide(
    data = ruuvi.data,
    timewindow_min = 5,
    tag_name = FALSE
   )
 }
}
