% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruuvifunctions.R
\name{PlotRuuvi}
\alias{PlotRuuvi}
\title{PlotRuuvi}
\usage{
PlotRuuvi(data)
}
\arguments{
\item{pathRuuvitag}{local path to csv files}
}
\description{
plot ruuvi data
}
\examples{
 \dontrun{

   data.wide <- MakeDataWide(
    data = ruuvi.data,
    timewindow_min = 5,
    tag_name = FALSE
   )

   PlotRuuvi(data = data.wide)
 }
}
