% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruuvifunctions.R
\name{SaveRuuviAsJson}
\alias{SaveRuuviAsJson}
\title{read ruuvi data from local machine}
\usage{
SaveRuuviAsJson(data, pathRuuviExport = NULL, filenamejson = NULL)
}
\arguments{
\item{data}{data}

\item{pathRuuviExport}{path to json file}

\item{filenamejson}{clean.field.names}
}
\value{
This function returns N ruuvi exports (in the selected folder) as one data frame object.
}
\description{
SaveRuuviAsJson data as json
}
\examples{
 \dontrun{

  p <- SetupRuuviR(p = NULL, username = "talonendm", printparameters = TRUE)

  rd1 <- LoadRuuviExports(
     pathRuuvitag = 'C:/Users/talon_000/Dropbox/2021/ruuvi/m210829',
     clean.field.names = FALSE
  )

  rd2 <- RuuviR::MakeDataWide(data = rd1)

  SaveRuuviAsJson(data = rd2,
   pathRuuviExport = p$path.json.export,
   filenamejson = p$ruuvi.datetag
  )

  # https://dl.dropboxusercontent.com/s/ywk9qitmuq5digf/m210829.json?dl=0

 }
}
