% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruuvifunctions.R
\name{SetupRuuviR}
\alias{SetupRuuviR}
\title{setup}
\usage{
SetupRuuviR(
  pathRuuviExport = "",
  username = "talonendm",
  printparameters = FALSE
)
}
\arguments{
\item{username}{username}

\item{printparameters}{printparameters}

\item{p}{is NULL}
}
\value{
This function creates parameter object p for development.
}
\description{
SetupRuuviR parameter object p
}
\examples{
 \dontrun{
  p <- SetupRuuviR(p = NULL, username = "talonendm", printparameters = TRUE)
 }
}
