% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruuvifunctions.R
\name{LoadRuuviExports}
\alias{LoadRuuviExports}
\title{read ruuvi data from local machine}
\usage{
LoadRuuviExports(pathRuuvitag = "")
}
\arguments{
\item{pathRuuvitag}{local path to csv files}
}
\value{
This function returns N ruuvi exports (in the selected folder) as one data frame object.
}
\description{
csv data to R
}
\examples{
 \dontrun{
  ruuvi.data <- LoadRuuviExports(
   pathRuuvitag = 'C:/Users/talon_000/Dropbox/2021/ruuvi/m210829'
  )
 }
}
